+++ lib/model/FunctionsType.d.ts
@@ -1,4 +1,4 @@
 export declare type FunctionsType = {
-    [name: string]: Function;
-};
+  [name: string]: Function
+}
 //# sourceMappingURL=FunctionsType.d.ts.map